rename Singleton.py to lowercase
