+++ src/benchmark/frameworks/singleton.py
